+++ code/renderer/graphics_api/texture.hpp
@@ -13,11 +13,11 @@ struct TextureCreateInfo
     ImageType type       = ImageType::TYPE_2D;
     PixelFormat format   = PixelFormat::NUM_PIXEL_FORMATS;
     SamplerType sampler  = SAMPLER_TRILINEAR_WRAP_U_WRAP_V;
-    uint32_t mipLevels   = 1;
-    uint32_t arrayLayers = 1;
-    uint32_t width       = 0;
-    uint32_t height      = 0;
-    uint32_t depth       = 1;
+    uint8_t mipLevels    = 1;
+    uint16_t arrayLayers = 1;
+    uint16_t width       = 0;
+    uint16_t height      = 0;
+    uint16_t depth       = 1;
     VkImageUsageFlags usage =
         VK_IMAGE_USAGE_SAMPLED_BIT | VK_IMAGE_USAGE_STORAGE_BIT | VK_IMAGE_USAGE_TRANSFER_SRC_BIT | VK_IMAGE_USAGE_TRANSFER_DST_BIT;
